Move ami_cleanup from install to platform cookbook
