+++ cookbooks/aws-parallelcluster-install/recipes/default.rb
@@ -1,0 +1,1 @@
+# Required for ChefSpec tests to be able to converge
