+++ cookbooks/aws-parallelcluster-install/spec/unit/recipes/fake_spec.rb
@@ -1,0 +1,12 @@
+require 'spec_helper'
+
+# Without this fake test other tests don't work
+# Leaving here for the time being, until the entire cookbook is removed
+describe 'aws-parallelcluster-install' do
+  cached(:chef_run) do
+    ChefSpec::Runner.new.converge(described_recipe)
+  end
+  it 'Does nothing' do
+    is_expected.not_to install_package("any")
+  end
+end
