+++ cookbooks/aws-parallelcluster-platform/files/ami_cleanup.sh
