+++ kitchen.recipes-install.yml
@@ -99,12 +99,6 @@ suites:
         head_node_private_ip: '127.0.0.1'
         head_node_home_path: '/fake_headnode_home'
         shared_dir_head: '/fake_headnode_shared'
-  - name: ami_cleanup
-    run_list:
-      - recipe[aws-parallelcluster-install::ami_cleanup]
-    verifier:
-      controls:
-        - ami_cleanup_file_created
   - name: pmix_installed
     run_list:
       - recipe[aws-parallelcluster-tests::setup]
